+++ tests/Unit/EventTest.php
@@ -201,7 +201,7 @@ class EventTest extends TestCase
     /**
      * @test
      * @group legacy
-     * @expectedDeprecation Using cron expression with more than 5 parts is deprecated from v1.8 and will result in exception in v2.0. If you are using dragonmantank/cron-expression package be aware that passing more than five parts to this method will result in exception.
+     * @expectedDeprecation Using cron expression with more than 5 parts is deprecated from v1.9 and will result in exception in v2.0. If you are using dragonmantank/cron-expression package be aware that passing more than five parts to this method will result in exception.
      */
     public function moreThanFivePartsInCronExpressionResultsInDeprecationNotice()
     {